+++ src/components/ImportSpatialIllusionsStep.vue
@@ -10,7 +10,20 @@
             v-model="parentUnitId"
           />
         </section>
-        <section class="mt-4"></section>
+        <section class="mt-4">
+          <DataGrid
+            :data="[props.data]"
+            :columns="computedColumns"
+            :row-height="40"
+            class="max-h-[40vh]"
+            :initial-state="{ expanded: true }"
+            :get-sub-rows="(row) => row.subOrganizations"
+          />
+        </section>
+        <section class="mt-4 flex gap-4 p-1">
+          <InputCheckbox v-model="useFillColor" label="Use custom fill color" />
+          <InputCheckbox v-model="expandedStackedUnits" label="Expand stacked units" />
+        </section>
       </div>
 
       <footer class="flex flex-shrink-0 items-center justify-end space-x-2 pt-4">
@@ -22,10 +35,8 @@
 </template>
 
 <script setup lang="ts">
-import { computed, ref } from "vue";
+import { computed, h, ref } from "vue";
 import BaseButton from "@/components/BaseButton.vue";
-import { useNotifications } from "@/composables/notifications";
-import { useImportStore } from "@/stores/importExportStore";
 import { injectStrict } from "@/utils";
 import { activeScenarioKey } from "@/components/injects";
 import { NUnit, NUnitAdd } from "@/types/internalModels";
@@ -38,6 +49,10 @@ import {
   SpatialIllusionsOrbat,
 } from "@/types/externalModels";
 import { EntityId } from "@/types/base";
+import { CellContext, ColumnDef } from "@tanstack/vue-table";
+import DataGrid from "@/modules/grid/DataGrid.vue";
+import OrbatCellRenderer from "@/components/OrbatCellRenderer.vue";
+import InputCheckbox from "@/components/InputCheckbox.vue";
 
 interface Props {
   data: SpatialIllusionsOrbat;
@@ -47,6 +62,9 @@ const props = defineProps<Props>();
 const emit = defineEmits(["cancel", "loaded"]);
 const { unitActions, store: scnStore, time } = injectStrict(activeScenarioKey);
 const { state } = scnStore;
+
+const useFillColor = ref(true);
+const expandedStackedUnits = ref(true);
 
 const rootUnitItems = computed((): SymbolItem[] => {
   return Object.values(state.sideGroupMap)
@@ -63,13 +81,68 @@ const rootUnitItems = computed((): SymbolItem[] => {
 
 const parentUnitId = ref(rootUnitItems.value[0].code as string);
 
+function renderExpandCell({ getValue, row }: CellContext<SpatialIllusionsOrbat, string>) {
+  let symbolOptions: Record<string, any> = {};
+  if (useFillColor.value && row.original.options.fillColor) {
+    symbolOptions["fillColor"] = row.original.options.fillColor;
+  }
+  return h(OrbatCellRenderer, {
+    value: getValue(),
+    sidc: row.original.options.sidc,
+    expanded: row.getIsExpanded(),
+    level: row.depth,
+    symbolOptions,
+  });
+}
+
+const computedColumns = computed((): (ColumnDef<SpatialIllusionsOrbat> | false)[] => {
+  return [
+    {
+      accessorFn: (f) => f.options.uniqueDesignation,
+      id: "name",
+      header: "Unit",
+      cell: renderExpandCell,
+      enableGlobalFilter: false,
+      size: 350,
+      enableSorting: false,
+    },
+    {
+      id: "additionalInformation",
+      header: "Additional information",
+      accessorFn: (f) => f.options.additionalInformation,
+      enableSorting: false,
+    },
+    {
+      id: "stack",
+      header: "Stack",
+      accessorFn: (f) => f.options.stack,
+      enableSorting: false,
+    },
+    {
+      id: "fillColor",
+      header: "Fill color",
+      accessorFn: (f) => f.options.fillColor,
+      enableSorting: false,
+    },
+
+    {
+      id: "reinforced",
+      header: "Reinforced",
+      accessorFn: (f) => f.options.reinforced,
+      enableSorting: false,
+    },
+  ];
+});
+
 async function onLoad(e: Event) {
   const { side } = unitActions.getUnitHierarchy(parentUnitId.value);
   const oob = props.data;
 
   scnStore.groupUpdate(() => {
     function helper(u: SpatialIllusionsOrbat, parentId: EntityId) {
-      const { uniqueDesignation: name = "", sidc, fillColor, stack = 1 } = u.options;
+      const { uniqueDesignation: name = "", sidc } = u.options;
+      const stack = expandedStackedUnits.value ? u.options.stack || 1 : 1;
+      const fillColor = useFillColor.value ? u.options.fillColor : undefined;
 
       const newUnit: NUnitAdd = {
         name,