+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 September 2024
 
+- Add preview and more options to Spatial Illusions unitgenerator ORBAT import.
 - Add support for transforming multiple features at once.
 - Add drag and drop from unit breadcrumbs to the map.
 - Add experimental support for time varying scenario feature geometries.